test: fix review fixes
