+++ packages/agent/test/agent/routes/modification/dissociate-delete-related.test.ts
@@ -1,15 +1,57 @@
-import { Aggregator, Filter, Operator, ValidationError } from '@forestadmin/datasource-toolkit';
+import {
+  Aggregator,
+  Filter,
+  Operator,
+  PrimitiveTypes,
+  ValidationError,
+} from '@forestadmin/datasource-toolkit';
 import { createMockContext } from '@shopify/jest-koa-mocks';
 
 import * as factories from '../../__factories__';
 import { HttpCode } from '../../../../src/agent/types';
-import {
-  setupWithManyToManyRelation,
-  setupWithOneToManyRelation,
-} from '../../__helper__/setup-with-relation';
 import DissociateDeleteRoute from '../../../../src/agent/routes/modification/dissociate-delete-related';
 
 describe('DissociateDeleteRelatedRoute', () => {
+  const setupWithOneToManyRelation = () => {
+    const services = factories.forestAdminHttpDriverServices.build();
+    const options = factories.forestAdminHttpDriverOptions.build();
+    const router = factories.router.mockAllMethods().build();
+
+    const bookPersons = factories.collection.build({
+      name: 'bookPersons',
+      schema: factories.collectionSchema.build({
+        fields: {
+          id: factories.columnSchema.isPrimaryKey().build(),
+          bookId: factories.columnSchema.build({
+            columnType: PrimitiveTypes.Uuid,
+          }),
+        },
+      }),
+    });
+
+    const books = factories.collection.build({
+      name: 'books',
+      schema: factories.collectionSchema.build({
+        fields: {
+          id: factories.columnSchema.isPrimaryKey().build(),
+          myBookPersons: factories.oneToManySchema.build({
+            foreignCollection: 'bookPersons',
+            originKey: 'bookId',
+            originKeyTarget: 'id',
+          }),
+        },
+      }),
+    });
+    const dataSource = factories.dataSource.buildWithCollections([bookPersons, books]);
+
+    return {
+      dataSource,
+      services,
+      options,
+      router,
+    };
+  };
+
   test('should register the private route', () => {
     const { services, dataSource, options, router } = setupWithOneToManyRelation();
 
@@ -30,6 +72,80 @@ describe('DissociateDeleteRelatedRoute', () => {
   });
 
   describe('handleDissociateDeleteRelatedRoute', () => {
+    const setupWithManyToManyRelation = () => {
+      const services = factories.forestAdminHttpDriverServices.build();
+      const options = factories.forestAdminHttpDriverOptions.build();
+      const router = factories.router.mockAllMethods().build();
+
+      const libraries = factories.collection.build({
+        name: 'libraries',
+        schema: factories.collectionSchema.build({
+          fields: {
+            id: factories.columnSchema.isPrimaryKey().build(),
+            manyToManyRelationField: factories.manyToManySchema.build({
+              throughCollection: 'librariesBooks',
+              foreignRelation: 'myBook',
+              foreignCollection: 'books',
+              foreignKey: 'bookId',
+              foreignKeyTarget: 'id',
+              originKey: 'libraryId',
+              originKeyTarget: 'id',
+            }),
+          },
+        }),
+      });
+
+      const librariesBooks = factories.collection.build({
+        name: 'librariesBooks',
+        schema: factories.collectionSchema.build({
+          fields: {
+            bookId: factories.columnSchema.isPrimaryKey().build(),
+            libraryId: factories.columnSchema.isPrimaryKey().build(),
+            myBook: factories.manyToOneSchema.build({
+              foreignCollection: 'books',
+              foreignKey: 'bookId',
+              foreignKeyTarget: 'id',
+            }),
+            myLibrary: factories.manyToOneSchema.build({
+              foreignCollection: 'libraries',
+              foreignKey: 'libraryId',
+              foreignKeyTarget: 'id',
+            }),
+          },
+        }),
+      });
+
+      const books = factories.collection.build({
+        name: 'books',
+        schema: factories.collectionSchema.build({
+          fields: {
+            id: factories.columnSchema.isPrimaryKey().build(),
+            manyToManyRelationField: factories.manyToManySchema.build({
+              throughCollection: 'librariesBooks',
+              foreignRelation: 'myLibrary',
+              foreignCollection: 'libraries',
+              foreignKey: 'libraryId',
+              foreignKeyTarget: 'id',
+              originKey: 'bookId',
+              originKeyTarget: 'id',
+            }),
+          },
+        }),
+      });
+      const dataSource = factories.dataSource.buildWithCollections([
+        librariesBooks,
+        books,
+        libraries,
+      ]);
+
+      return {
+        dataSource,
+        services,
+        options,
+        router,
+      };
+    };
+
     test('should throw an error when an empty id list is passed', async () => {
       const { services, dataSource, options } = setupWithManyToManyRelation();
       const count = new DissociateDeleteRoute(
